% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spFW_example_data.R
\docType{data}
\name{spFW_example_data}
\alias{spFW_example_data}
\title{Example Data for Analysis}
\format{A data frame with 5 elements, which are:
\describe{
\item{yield}{Plant yield}
\item{geno}{Plant genotype ID}
\item{loc}{Plant environment ID}
\item{range}{Plant x-coordinate at a given environment}
\item{pass}{Plant y-coordinate at a given environment}
}}
\usage{
spFW_example_data
}
\description{
A data frame containing the plant yield, genotypes, environments, and within-field positions.
}
\keyword{datasets}
